change long to int

+++ TweetService/Infrastructure/ITweetRepository.cs
@@ -4,8 +4,8 @@ namespace TweetService.Infrastructure;
 
 public interface ITweetRepository
 {
-    public IEnumerable<Tweet> AllFrom(ulong id);
-    public IEnumerable<Tweet> AllRecent(IEnumerable<ulong> ids, int fromUtc, int toUtc);
+    public IEnumerable<Tweet> AllFrom(int id);
+    public IEnumerable<Tweet> AllRecent(IEnumerable<int> ids, int fromUtc, int toUtc);
     public Tweet Create(Tweet tweet);
     public bool Delete(int id);
     public Tweet Single(int id);
